+++ src/tabicl/prior/graph_lib/_function.py
