+++ src/tabicl/prior/graph_lib/_graph_function.py
