+++ src/tabicl/prior/graph_lib/_matrix.py
