+++ src/tabicl/prior/graph_lib/_multi_function.py
