+++ src/tabicl/prior/graph_lib/_node_function.py
