+++ src/tabicl/prior/graph_lib/_properties.py
