+++ src/tabicl/prior/graph_lib/_weights.py
